Friend requests implementation and navbar.;
Implemented navbar for easier user navigation of the website and completed friend requesting system. Fellow user/dog search is a next possibility. QR code scanning was initially planned, but JS would be required for this to be an easier implementation to my understanding. User or dog searching may be explored or a user specific web route to initiate friend request (similar to what a user might see on for a social media channel like Discord where the user is taken to a website where they can accept an invite to the channel, then they are added.