Query fixing and friend request modification.
Fixed queries to properly display user's own dogs and friends' dogs. Addition of add_friend template in case I want to create a routing system for users to get rerouted to accepting friend request similar to discord implementation. In the meantime, a user can copy the link to their user_profile that they can send to friends to add them (this is available on their own user_profile).